12-01-20-adm

+++ rev_uso_fun.php
@@ -5,10 +5,11 @@ header("Location:index.php");
 die;
 }
 include_once "ConAL.php";
-$ida = filter_input(INPUT_POST,'ida',FILTER_SANITIZE_STRING);
-
-$vl ="DELETE FROM log WHERE log.id =".$ida;
-$rvl = mysqli_query($conn, $vl);
-$_SESSION['ifon']="<script>alert('Removido com sucesso')</script>";
-header("Location:rev_uso.php");
+if(is_numeric($_GET['id'])){
+    $ida = $_GET['id'];
+    $vl ="DELETE FROM log WHERE log.id =".$ida;
+    $rvl = mysqli_query($conn, $vl);
+    $_SESSION['ifon']="<script>alert('Removido com sucesso')</script>";
+    header("Location:alter_uso.php");
+}
  ?>
